requirements.txt
torch==2.3.1
scikit-learn==1.5.1
matplotlib==3.9.1
torchvision==0.18.1
pandas==2.2.2 
click==8.1.7
PyYAML==6.0.1
torch-summary==1.4.5